super_scan
git-svn-id: https://svn.nmap.org/trunk/nmap@289 e0a8ed71-7df4-0310-8962-fdc924857419

+++ nmap.c
@@ -1916,7 +1916,7 @@ if (o.debugging || o.verbose)
 	}
       }
 
-      printf("Ideal number of queries: %d\n", numqueries_ideal);
+      printf("Ideal number of queries: %d\n", (int) numqueries_ideal);
       /* Now that we have sent the packets we wait for responses */
       while (( ip = (struct ip*) readip_pcap(pd, &bytes))) {
 	if (bytes < (4 * ip->ip_hl) + 4)
